modified:   src/controllers/authController.js 	modified:   src/controllers/notesController.js 	modified:   src/models/note.js

+++ src/controllers/notesController.js
@@ -11,7 +11,7 @@ export const getAllNotes = async (req, res) => {
   const skip = (page - 1) * perPage;
 
   // Створюємо базовий запит до колекції
-  const notesQuery = Note.find();
+  const notesQuery = Note.find({userId: req.user._id});
 
   // Будуємо фільтр
   if (tag) {
@@ -45,7 +45,7 @@ export const getAllNotes = async (req, res) => {
 // Опис контролера getNoteById - отримати одну нотатку за id або помилку 404
 export const getNoteById = async (req, res, next) => {
   const { noteId } = req.params;
-  const note = await Note.findById(noteId);
+  const note = await Note.findOne({ _id: noteId, userId: req.user._id });
 
   if (!note) {
     // замість «звичайної» помилки, явно вказуємо код 404 і повідомлення
@@ -60,7 +60,7 @@ export const getNoteById = async (req, res, next) => {
 // Додаємо можливість створювати нову нотатку до колекції за маршрутом POST /notes.
 // У запиті будемо очікувати тіло запиту, яке приходитиме як JSON.
 export const createNote = async (req, res) => {
-  const note = await Note.create(req.body);
+  const note = await Note.create({ ...req.body, userId: req.user._id });
   res.status(201).json(note);
 };
 
@@ -72,7 +72,7 @@ export const createNote = async (req, res) => {
 // Опис контролера deleteNote - для видалення існуючої нотатки за її ідентифікатором
 export const deleteNote = async (req, res, next) => {
   const { noteId } = req.params;
-  const note = await Note.findOneAndDelete({ _id: noteId });
+  const note = await Note.findOneAndDelete({ _id: noteId, userId: req.user._id });
 
   if (!note) {
     next(createHttpError(404, 'Note not found'));
@@ -86,7 +86,7 @@ export const updateNote = async (req, res, next) => {
   const { noteId } = req.params;
 
   const note = await Note.findOneAndUpdate(
-    { _id: noteId }, // Шукаємо по id
+    { _id: noteId, userId: req.user._id }, // Шукаємо по id
     req.body,
     { new: true }, // повертаємо оновлений документ
   );
